refactor(bench): extract model-agnostic kernel-bench harness into pegainfer-bench
Every model crate's kernel-report tooling re-implemented the same CUDA-event
timing loop, LatencyStats, KernelCall accessors, and device-alloc helpers. Lift
them into a new `pegainfer-bench` crate and migrate the two existing consumers.

- kimi-k2 `kernel_report.rs`: drop the local harness, re-export the shared types
  so the report bins are untouched (803 -> 663 lines). Non-optional dep because
  the harness is an always-compiled lib module.
- qwen3-4b `qwen3_model_report.rs`: adopt the shared LatencyStats/accessors;
  keep its cold-L2 `measure_loop` and attr-filtered `bench_key` local because
  their measurement semantics deliberately differ from kimi's warm-cache /
  raw-key versions (1079 -> 981 lines). Optional dep gated on `kernel-report`
  since the harness lives in the report bin.

Net -65 source lines; the real win is one canonical harness instead of two
copies that had already silently diverged (warm vs cold-L2 timing).

Verified: pegainfer-bench checks model-agnostic (Kimi kernels disabled); kimi +
qwen3 report bins compile under `kernel-report`; kimi 18 / qwen3 17 host unit
tests pass; no-feature qwen3 build does not pull pegainfer-bench.
