+++ pegainfer-qwen3-4b/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 autobenches = false
 
 [dependencies]
+pegainfer-bench = { workspace = true, optional = true }
 pegainfer-core = { workspace = true }
 pegainfer-cupti = { workspace = true, optional = true }
 pegainfer-kernels = { workspace = true }
@@ -32,6 +33,7 @@ kernel-report = [
     "dep:clap",
     "dep:comfy-table",
     "dep:hex",
+    "dep:pegainfer-bench",
     "dep:pegainfer-cupti",
     "dep:sha2",
     "dep:toml",
